requirements.txt
numpy
scikit-learn
fastapi
uvicorn[standard]
python-multipart
fastapi-cors
pymupdf
python-docx